renaming files
renamed main.py -> console_game.py and app.py -> main.py so that it fits with the pygbag naming scheme
